requirements.txt
spotipy
python-dotenv
pandas
numpy